feat: 重构 ChatMessageBubble 测试，新增测试文件并确保样式正常渲染
